refactor: remove paste HTML handling configuration and related tests
